add non-recursive lock semantics
Add support for non-recursive locks.  By default, all lockfiles are
recursive.  A non-recursive lock can now be created via `l :=
lockfile.New("/example/path.lock", false)` as the second parameter
indicates if the lock is recursive or not.  Notice that the second
parameter is optional (default=true), which is required to remain
backwards compatible with existing users.

Signed-off-by: Valentin Rothberg <vrothberg@suse.com>